+++ src/app/_dto/misskey-callback/callback-token-claim.dto.ts
@@ -1,0 +1,13 @@
+import { IsString } from "class-validator";
+
+export class callbackTokenClaimPayload {
+    //Misskey Access key 를 요청할 misskey의 host (예: serafuku.moe)
+    @IsString()
+    misskeyHost: string;
+  
+    //callback으로 받은 토큰
+    @IsString()
+    callback_token: string;
+  }
+  
+
