+++ src/app/_dto/misskey-callback/user-info.dto.ts
@@ -1,0 +1,5 @@
+import type { MiUser } from "@/app/api/misskey-entities/user";
+
+export class userInfoPayload {
+  user: MiUser;
+}
