+++ src/app/misskey-callback/page.tsx
@@ -5,7 +5,8 @@ import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { login } from "./actions";
 import { MiUser as MiUser } from "../api/misskey-entities/user";
-import { callbackTokenClaimPayload, userInfoPayload } from "..";
+import { callbackTokenClaimPayload } from "../_dto/misskey-callback/callback-token-claim.dto";
+import { userInfoPayload } from "../_dto/misskey-callback/user-info.dto";
 
 export type DBpayload = {
   account: string;
